Move validateUsername to dummy
